subtask Controller session

+++ src/main/java/com/example/projectpeak1/controller/SubtaskController.java
@@ -34,25 +34,30 @@ public class SubtaskController {
     private int getUserId(HttpSession session) {
         return (int) session.getAttribute("userId");
     }
+    private boolean isLoggedIn(HttpSession session){
+        return session.getAttribute("userId") != null;
+    }
 
     @GetMapping("/createSubtask/{id}")
     public String createTask(@PathVariable("id") int taskId, HttpSession session, Model model) {
-        int userId = getUserId(session);
-        if (userId == 0) {               // Check that the session with userid is logged in.
-            return "login_HTML/login";
+        if (!isLoggedIn(session)) {
+            return "redirect:/login";
         }
+
         TaskAndSubtaskDTO taskAndSubtask = subtaskService.getTaskAndSubTaskById(taskId);
         model.addAttribute("task", taskAndSubtask);
 
         if(!isUserAuthorized(session, taskAndSubtask.getProjectId())){
             return "error_HTML/accessDenied";
         }
-
-        User user = subtaskService.getUserFromId(userId);
-        model.addAttribute("user", user);
-        model.addAttribute("taskId", taskId);
-        model.addAttribute("subtask", new Subtask());
-        return "subtask_HTML/addSubtask";
+        else {
+            int userId = getUserId(session);
+            User user = subtaskService.getUserFromId(userId);
+            model.addAttribute("user", user);
+            model.addAttribute("taskId", taskId);
+            model.addAttribute("subtask", new Subtask());
+            return "subtask_HTML/addSubtask";
+        }
     }
 
     @PostMapping(value = {"/createSubtask/{id}"})
@@ -65,26 +70,27 @@ public class SubtaskController {
 
     @GetMapping("/showSubtask/{taskId}")
     public String showSubtaskDetails(HttpSession session, @PathVariable("taskId") int taskId, Model model) {
-        int userId = getUserId(session);
-        if (userId == 0) {
-            return "login_HTML/login";
+        if (!isLoggedIn(session)) {
+            return "redirect:/login";
         }
-        User user = subtaskService.getUserFromId(userId);
-        model.addAttribute("user", user);
-        TaskAndSubtaskDTO taskAndSubtask = subtaskService.getTaskAndSubTaskById(taskId);
-        model.addAttribute("taskAndSubtask", taskAndSubtask);
+        else {
+            int userId = getUserId(session);
+            User user = subtaskService.getUserFromId(userId);
+            model.addAttribute("user", user);
+            TaskAndSubtaskDTO taskAndSubtask = subtaskService.getTaskAndSubTaskById(taskId);
+            model.addAttribute("taskAndSubtask", taskAndSubtask);
 
-        if(!isUserAuthorized(session, taskAndSubtask.getProjectId())){
-            return "error_HTML/accessDenied";
+            if(!isUserAuthorized(session, taskAndSubtask.getProjectId())){
+                return "error_HTML/accessDenied";
+            }
+            return "subtask_HTML/showSubtask";
         }
-        return "subtask_HTML/showSubtask";
     }
 
     @GetMapping("/editSubtask/{id}")
     public String editSubtask(HttpSession session, @PathVariable("id") int subtaskId, Model model) {
-        int userId = getUserId(session);
-        if (userId == 0) {
-            return "login_HTML/login";
+        if (!isLoggedIn(session)) {
+            return "redirect:/login";
         }
         Subtask subtask1 = subtaskService.getSubtaskById(subtaskId); //get the subtask, so we can show in html edit site.
         model.addAttribute("subtask", subtask1);
@@ -94,10 +100,13 @@ public class SubtaskController {
         if(!isUserAuthorized(session, taskAndSubtask.getProjectId())){
             return "error_HTML/accessDenied";
         }
-        User user = subtaskService.getUserFromId(userId);
-        model.addAttribute("user", user);
+        else {
+            int userId = getUserId(session);
+            User user = subtaskService.getUserFromId(userId);
+            model.addAttribute("user", user);
+            return "subtask_HTML/editSubtask";
+        }
 
-        return "subtask_HTML/editSubtask";
     }
 
     @PostMapping("/editSubtask")
@@ -108,47 +117,50 @@ public class SubtaskController {
 
     @GetMapping(value = "/deleteSubtask/{id}")
     public String deleteTask(HttpSession session, @PathVariable("id") int subtaskId) throws LoginException {
+        if (!isLoggedIn(session)) {
+            return "redirect:/login";
+        }
         int taskId = subtaskService.getTaskIdBySubtaskId(subtaskId);
         if(!isUserAuthorized(session, subtaskService.getProjectIdBtSubtaskId(subtaskId))){
             return "error_HTML/accessDenied";
         }
-        subtaskService.deleteSubtask(subtaskId);
-        return "redirect:/showSubtask/" + taskId;
+        else {
+            subtaskService.deleteSubtask(subtaskId);
+            return "redirect:/showSubtask/" + taskId;
+        }
     }
 
     @GetMapping(value = {"/doneSubtask/{id}"})
     public String doneSubtask(HttpSession session, @PathVariable("id") int id) throws LoginException {
-        int userId = getUserId(session);
-        if (userId == 0) {
-            return "login_HTML/login";
+        if (!isLoggedIn(session)) {
+            return "redirect:/login";
         }
         if(!isUserAuthorized(session, subtaskService.getProjectIdBtSubtaskId(id))){
             return "error_HTML/accessDenied";
         }
-        int taskId = subtaskService.getTaskIdBySubtaskId(id);
-        subtaskService.doneSubtask(id);
+        else {
+            int taskId = subtaskService.getTaskIdBySubtaskId(id);
+            subtaskService.doneSubtask(id);
+            return "redirect:/showSubtask/" + taskId;
+        }
 
-
-        return "redirect:/showSubtask/" + taskId;
 
     }
 
     @GetMapping(value = {"/showAllDoneSubtask/{taskId}"})
     public String seeAllDoneSubtask(HttpSession session, Model model, @PathVariable int taskId) {
-        int userId = getUserId(session);
-        if (userId == 0) {
-            return "login_HTML/login";
+        if (!isLoggedIn(session)) {
+            return "redirect:/login";
         }
-
-
-        User user = subtaskService.getUserFromId(userId);
-        model.addAttribute("user", user);
-
-        List<DoneSubtaskDTO> doneSubtaskDTOS = subtaskService.getAllDoneSubtask(taskId);
-        model.addAttribute("seeDoneSubtask", doneSubtaskDTOS);
-        model.addAttribute("taskId", taskId);
-
-        return "subtask_HTML/showAllDoneSubtask";
+        else {
+            int userId = getUserId(session);
+            User user = subtaskService.getUserFromId(userId);
+            model.addAttribute("user", user);
+            List<DoneSubtaskDTO> doneSubtaskDTOS = subtaskService.getAllDoneSubtask(taskId);
+            model.addAttribute("seeDoneSubtask", doneSubtaskDTOS);
+            model.addAttribute("taskId", taskId);
+            return "subtask_HTML/showAllDoneSubtask";
+        }
 
     }
